Minor refactor to TrapManager.
